+++ custom_components/lutarym_ppc_smgw/api.py
@@ -1,4 +1,4 @@
-# Integrationsversion: 2.4.4
+# Integrationsversion: 2.4.5
 """Client für die HAN-Schnittstelle eines PPC LTE Smart Meter Gateways (SMGW).
 
 Das Gateway bietet keine "echte" REST-API, sondern nur das interne